修改了提测通过率不能是小数的bug

+++ dal/src/main/java/com/timevale/forward/dal/entity/TestBillDO.java
@@ -49,7 +49,7 @@ public class TestBillDO extends BaseDO {
     /**
      * 提测通过率
      */
-    private Integer passRate;
+    private Double passRate;
     /**
      * 影响范围与变更SQL
      */
